rename module for the same style
