+++ pyCADD/Demand/__init__.py
@@ -1,0 +1,1 @@
+from pyCADD.Demand.core import QueryClient
