+++ pyCADD/Demand/__main__.py
@@ -1,4 +1,4 @@
-from pyCADD.query.cli import main
+from pyCADD.Demand.cli import main
 
 if __name__ == '__main__':
     main()
