--- pyCADD/query/__init__.py
@@ -1,2 +1,0 @@
-
-from pyCADD.query.core import QueryClient
